requirements.txt
flask==2.3.3
jinja2==3.1.2
werkzeug==2.3.7
langdetect==1.0.9
transformers==4.35.2
torch==2.8.0
google-generativeai==0.3.1
pandas==2.0.3
pytest==7.4.3
pytest-flask==1.3.0
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0